Revert "Mapped kernel at 3Gb from the begining"
This reverts commit 652a76a92139265a49a75e34ea9f60ba20ca1fad.

+++ include/com.h
@@ -4,10 +4,10 @@
 #include "defs.h"
 
 
-void com_putc(size_t port, int s);
-void com_puts(size_t port, char *s);
-int com_getc(size_t port);
-void com_init(size_t port);
+void com_putc(uint16_t port, int s);
+void com_puts(uint16_t port, char *s);
+int com_getc(uint16_t port);
+void com_init(uint16_t port);
 
 
 #endif /* _COM_H_ */
